Enabled auth in archive

+++ api/archive/api.py
@@ -5,14 +5,17 @@ import archive.schemas
 
 from typing import List
 
+from security.api import AuthBearer
+
 router = Router()
 
-@router.post('/post_record')
+@router.post('/post_record', auth=AuthBearer())
 def post_archive_record(request, r: archive.schemas.ArchiveRecordIn):
     try:
         if r.note.__len__() < 4:
             raise ValueError('Note must contain at least 4 chars')
         record = r.get_object()
+        record.owner_token = request.auth.owner_token
         record.save()
     except Exception as e:
         print(f'Failed to post record: {e}')
@@ -21,13 +24,13 @@ def post_archive_record(request, r: archive.schemas.ArchiveRecordIn):
         return {'result': 'success', 'code': 0}
 
 
-@router.get('/get_record', response=archive.schemas.ArchiveRecordOut)
+@router.get('/get_record', response=archive.schemas.ArchiveRecordOut, auth=AuthBearer())
 def get_archive_record(request, id: int):
     try:
         if id <= 0:
             raise ValueError('Bad id')
         rs = records.models.Record.objects.filter(
-            id=id, record_type=records.models.RecordTypes.ARCHIVE)
+            id=id, record_type=records.models.RecordTypes.ARCHIVE, owner_token=request.auth.owner_token)
         r = rs[0]
         return r
     except Exception as e:
@@ -35,19 +38,19 @@ def get_archive_record(request, id: int):
         return HttpResponseBadRequest(content=f'{e}')
 
 
-@router.get('/get_records', response=List[archive.schemas.ArchiveRecordOut])
+@router.get('/get_records', response=List[archive.schemas.ArchiveRecordOut], auth=AuthBearer())
 def get_archive_records(request, limit: int = 100, offset: int = 0):
     if limit > 500:
         limit = 500
     if offset < 0:
         offset = 0
-    return list(records.models.Record.objects.filter(record_type=records.models.RecordTypes.ARCHIVE).order_by('-id')[offset:(offset+limit)])
+    return list(records.models.Record.objects.filter(record_type=records.models.RecordTypes.ARCHIVE, owner_token=request.auth.owner_token).order_by('-id')[offset:(offset+limit)])
 
 
-@router.delete('/delete_record')
+@router.delete('/delete_record', auth=AuthBearer())
 def delete_archive_record(request, id: int):
     try:
-        ret = records.models.Record.objects.filter(id=id, record_type=records.models.RecordTypes.ARCHIVE).delete()
+        ret = records.models.Record.objects.filter(id=id, record_type=records.models.RecordTypes.ARCHIVE, owner_token=request.auth.owner_token).delete()
         if ret[0] == 0:
             raise ValueError('Deleted zero elements')
     except Exception as e:
@@ -57,13 +60,13 @@ def delete_archive_record(request, id: int):
         return {'result': 'success', 'code': 0}
 
 
-@router.patch('/make_archive')
+@router.patch('/make_archive', auth=AuthBearer())
 def archive_crate_record(request, id: int):
     try:
         if id <= 0:
             raise ValueError('Bad id')
         ret = records.models.Record.objects.filter(
-            id=id).update(record_type=records.models.RecordTypes.ARCHIVE)
+            id=id).update(record_type=records.models.RecordTypes.ARCHIVE, owner_token=request.auth.owner_token)
         print(ret)
         if ret == 0:
             raise ValueError('Changed zero elements')
